refactor(CartBtn): move component to feature folder
