+++ src/pages/MyFavorites.tsx
@@ -14,7 +14,7 @@ import {
 } from "@ionic/react";
 import { collection, getFirestore, query, where } from "firebase/firestore";
 
-import CartBtn from "../components/CartBtn";
+import CartBtn from "../features/Cart/components/CartBtn";
 import ProductCard from "../components/ProductCard";
 import { ProductConvert } from "../converters/products";
 import { ProductLoading } from "../constants";
